+++ src/Fantomas.Core/AstTransformer.fs
@@ -1236,7 +1236,7 @@ and visitSynPat (sp: SynPat) : TriviaNode =
 and visitSynConstructorArgs (ctorArgs: SynArgPats) : TriviaNode list =
     match ctorArgs with
     | SynArgPats.Pats pats -> List.map visitSynPat pats
-    | SynArgPats.NamePatPairs (pats, range) ->
+    | SynArgPats.NamePatPairs (pats, _, { ParenRange = StartEndRange 1 (lpr, range, rpr) }) ->
         let children =
             pats
             |> List.map (fun (ident, eqRange, pat) ->
@@ -1247,7 +1247,13 @@ and visitSynConstructorArgs (ctorArgs: SynArgPats) : TriviaNode list =
                        mkNode SynArgPats_NamePatPairs_Equals eqRange
                        visitSynPat pat |])
 
-        [ mkNodeWithChildren SynArgPats_NamePatPairs range (sortChildren [| yield! children |]) ]
+        [ mkNodeWithChildren
+              SynArgPats_NamePatPairs
+              range
+              (sortChildren
+                  [| yield mkNode SynArgPats_NamePatPairs_OpeningParenthesis lpr
+                     yield! children
+                     yield mkNode SynArgPats_NamePatPairs_ClosingParenthesis rpr |]) ]
 
 and visitSynComponentInfo (sci: SynComponentInfo) : TriviaNode list =
     match sci with